+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `Authorization` module with Casbin
+- Add `EnforceScope` type and module
+- Add `Session.Authorize.withScope` function
+- Add `JWT.TokenData` field `Subject`
 
 ## 8.0.0 - 2026-01-14
 - [**BC**] Rename